chore: Update Apache Airflow and Apache Spark dependencies
This commit updates the `pyproject.toml` and `requirements.txt` files to include the latest versions of the `apache-airflow` and `apache-airflow-providers-apache-spark` dependencies. Additionally, the `Dockerfile` has been modified to install OpenJDK 11 and set the `JAVA_HOME` environment variable. The new versions and dependencies will improve the performance and compatibility of the project.

New files:
- `astro/dags/first-dag.py`
- `astro/dags/second-dag.py`
- `README.md`

Modified files:
- `pyproject.toml`
- `astro/requirements.txt`
- `astro/Dockerfile`
- `spark-cluster/docker-compose.yaml`